storage: Introduce VolumeManifest.operation context
When performing datapath operations on a volume (eg. copying data) we
mark the volume ILLEGAL before starting the operation and only mark the
volume LEGAL again once the operation is finished.  As long as this is
all done with the volume lease held the engine can poll the volume from
any host to determine if the operation is running (lease held) and can
detect an interrupted/failed operation (lease free and volume ILLEGAL).

Later this contextmanager will be expanded to support volume generation
incrementation when exiting successfully which will also allow engine to
determine if an operation was completed successfully.

All metadata updates must be performed to a single block with one write
in order to ensure atomicity.

Change-Id: I30a3ac2971411778d24e007aac9fcb3009edb4c4
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64362
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>