tests: Remove TODO for SDM.copy_data abort before copy test
The abort before copy test is meant to exercise the code which exits the
job early (and without marking the destination volume illegal) if the
job has been aborted in an extremely narrow race window.  It's a
difficult test to write correctly and would not be testing anything
particularly interesting.  Drop the TODO.

Change-Id: I26e724560e9cb485fd4766ebcc04f644ad6431e4
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64480
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI