storage: Add support for generation to VolumeMetadata
We would like to add support for a new generation key in the volume
metadata.  GEN is a monotonically increasing integer that can be used
to determine whether certain operations have been completed on a volume
and to prevent two hosts from sequentially performing the same operation
in the event of double scheduling.

The new GEN key is optional.  If we read metadata without this key, a
default value of 0 will be reported.

This patch supports only the manipulation of the generation within
the VolumeMetadata class.  The value will be made available via
Volume.getInfo in a follow-up patch and will be changed by verbs in
subsequent patches.

Change-Id: Icb71e1fc78f6c1e411e725b26c48411ffd04d0b6
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64484
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI