StorageDomain.getInfo - report the first pv of the metadata lv
This patch adds the report of the first pv of the domain metadata lv
(for block domains) to StorageDomain.getInfo().
Currently we assume that the metadata lv is created on the first extant
of the device and verify that when performing different operations on
the domain.

As we intend to add support for operations that modify the vg
stracture (like reducing devices) - we want to prevent those
operations from being made on the first device of the metadata lv
as they might cause the metadata lv first extant to be changed from
0.

This patch adds the reporting of that info, so that the engine can
leverage the info to block such operations.

Change-Id: I32c847ae89b9f8f512c3dd8a0fff96fbc753ee5b
Signed-off-by: Liron Aravot <laravot@redhat.com>