virt: Try to detect non guest iniated shutdowns
When a host system shutsdown due to a variety of reasons (UPS, fencing...)
VMs running on that system get stopped before VDSM and libvirtd services
are stopped. This is due to the way how they are registerd with machined.

This results in the qemu processes being signaled with SIG_TERM and them
exiting. To libvirt this looks the same as if the shutdown would have been
initiated from within the VM and therefore emits a lifecycle event that
looks to VDSM as if the VM was stopped from within. VDSM in this case
reports the exit reason as user shutdown. This is a problem for HA VMs
which will not be automatically rescheduled when they see a user shutdown
exit reason.

This patch attempts to detect non user shutdowns. This requires the VM to
have the ovirt guest agent running to detect these scenarios properly.

Backport-To: 4.0
Bug-Url: https://bugzilla.redhat.com/1389332
Change-Id: Ie04b9806fbf0a81dc576aa28cfdda5edb079ce29
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64991
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65940