resourceManager: Add module api
resourceManager has a Java like verbose api that make it hard to use and
is not safe. Add module api hiding the single ResourceManager instance.

Instead of this:

    manager = rm.ResourceManager.getInstance()
    manager.acquireResource("mamespace", "name", rm.SHRARED)

Or this wrong usage:

    manager = ResourceManager()
    manager.acquireResource("mamespace", "name", rm.SHRARED)

We can use now this:

    rm.acquireResource("mamespace", "name", rm.SHRARED)

This is the simplest and safest way to implement a singleton in Python,
a module.

This patch adds the new api and update the tests to use this api.

Change-Id: Ie315f10bacdf49f1023d1991811afec28ed0d7f9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>