resourcemanager: Use new module api
Replace client code to use the resourceManager module directly instead
of using ResourceManager.getInstance().

After this patch only the resourceManager and resourceFactories are
using the old getInstance api.

Tests that used to mock modules "rmanager" variable are mocking the "rm"
variable, and storagefakelib.FakeResourceManager implements now the
resourceManager module api instead of the
resourceManager.ResourceManager class. This seems to be the quickest
solution for keeping these tests working, but we may use different way
later.

Change-Id: Ib33d9fad90551a0c2e9b1e8a89c5f9f920681c6b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65035
Reviewed-by: Adam Litke <alitke@redhat.com>