sp: Do not use rm.acquireResource return value
resourceManager.acquireResource is returning now a Resource object,
which can be used as a context manager. This usage is deprecated, as we
want to replace resourceManager with something much simpler.

Replace such code with usage of rm.acquireResource as a context manger.
This usage will be the only usage allowed in the future.

Change-Id: I8ff5653f0b15861d2034011f7fccc76bf25788f6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>