sp: Streamline building domains list for upgrade
Instead of creating a list of domain uuids, and removing the master
domain uuid, remove the master the domain uuid from the domains dict.
This avoids noisy exception handler and using \ continuation.

While touching this code, also fix the log, we are not marking *all*
domains for upgrade but only active domains.

Change-Id: I2be0a4816d733fdae13bcb933201b1ede795ca68
Signed-off-by: Nir Soffer <nsoffer@redhat.com>