sp: Streamline building domains list for upgrade
Instead of creating a list of domain uuids, and removing the master
domain uuid, remove the master the domain uuid from the domains dict.
This avoids noisy exception handler and using \ continuation.

While touching this code, also fix the log, we are not marking *all*
domains for upgrade but only active domains.

Change-Id: I2be0a4816d733fdae13bcb933201b1ede795ca68
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65043
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>