sp: Remove vol_extend_policy option
We had a vol_extend_policy option disabling the mailbox. This option is
not useful for anything but causing vms using thin block disks to pause
without a way to recover.

This option was added in commit 73adfc489e6e (Add spmProtect, release
lock and toggle lock functionality.  Add toggle lvextend functionality)
in 2009. I did not find why this option was added.

Change-Id: Ib067c09a985a633a452c476e21d8c2c073e6ca50
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65053
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>