vm: ignore more errors in isDomainReadyForCommands
Due to race on domain shutdown, and after migration end,
we may receive errors from libvirt which makes
the Vm.isDomainReadyForCommands() method fail with scary stacktrace.

This patch makes the method cope with VIR_ERR_OPERATION_UNSUPPORTED.
In the context of this method, this error can come only if the
domain is shutting down, so it is safe to swallow it in this case.

The real, big issue is how to properly handle the shutdown of a
domain. Currently, we have to deal with domains which are partially
shut down, so they are reported by libvirt and can return error
(like VIR_ERR_OPERATION_UNSUPPORTED) to methods which normally succeed.

One way to solve this is to make the shutdown (more) atomic:
either a domain is completely operative, or it is gone.
It is still to be decided if this could be done without libvirt changes.

Change-Id: I89ff61e0cd3bbb977833897c250ea337c86b9f80
Backport-To: 4.0
Bug-Url: https://bugzilla.redhat.com/1382578
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65131
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>