betterasynccore: Prevent double close
asyncore.dispatcher is defining a closing attribute, but never use it.
This make it possible to close a dispatcher multiple times, failing with
KeyError after the first call.

Because of the asynchronous nature of the reactor, it is very hard to
ensure that all dispatchers are closed before the reactor, so we need a
more robust close method.

This patch uses the closing attribute to eliminate double close errors.

Change-Id: Ia293deb79affc2f740d825e24ce8c66449385111
Signed-off-by: Nir Soffer <nsoffer@redhat.com>