net: edit nic detached from bridge but still attached to a vlan
We do not understand why, but the problem does not reproduce with dummy
devices. Therefore, we cannot easily test it. Trying to reproduce the
issue with a veth pair was deemed to much of a hassle. We test this
fix manually and do not include a regression test for it.

Change-Id: I460cb08cf436b932e7d9592557a03d7b6fc36a0e
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1372798
Reviewed-on: https://gerrit.ovirt.org/65231
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI