profiling: Decouple cpu profiling from vdsm config
Previously we accessed vdsm configuration to set module defaults during
import. This makes it impossible to use the module from another
application like supervdsm or ovirt-ha-agent.

The configuration is accessed now when starting an application wide
profile using the start() and stop() interface.

This change also simplify the tests, we can mock (almost) everything via
the module config.

vdsm configuration is still loaded when importing the profiler module,
more work is needed to avoid this.

Change-Id: I711c7947534841f77b88c6067c992ca1a3a42fdb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>