profiling: Add devel:cpu_profile_filename option
This is mainly for simplifying the tests. Now we can control the test
environment by mocking the config. Additionally, we use now a proper
temporary file for the saved profile, so the tests run now on a read
only file system.

This may also be useful for developers that want to change the location
of the application wide profile.

Change-Id: Id0a0800d08850fb236803813532210729ef16814
Signed-off-by: Nir Soffer <nsoffer@redhat.com>