properties: Make password test data more clear
The old test used the value b"12345678", which may confuse the
maintainer of this file about to think that this may be a Unicode
string. Replace with b"\x80\x81\x82\x83", which is clearly binary data,
and also will fail to decode to Unicode using many encodings.

Change-Id: Ifb507d5d728bc82a1698ed3191573f56f06323b0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65258
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>