build: Disable tests during build
Tests are needed for development, not for building a package. This
allows us to use latest and greatest development tools, which are not
available in brew or koji.

Since we install nose using pip, remove the build requires - we don't
need now nose to build rpms.

Also remove nose from the automation *.packages files, since we install
it from pip.

Finally remove the nose check from configure, since configure is run
during the build, and we cannot require nose at this point. We are
checking nose when running the tests.

We still install python 3 nose using the distribution packages, since it
is new enough, and mixing pip and pip3 on same machine does not work
well.

Change-Id: I9e3589c365166f934f117b53c65cea4b90db3516
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/63966
Reviewed-by: Edward Haas <edwardh@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>