utils: Wait for terminated process
utils.terminating was not waiting for a terminated process, passing the
process to zombiereaper. This behavior may be disastrous storage and
mostly unwanted for other code.

When we run a child process touching shared storage, we must wait for
the child process. If we kill the child process without waiting, engine
may start the same operation on another host, while the original child
process is still running, possibly in interruptible state, trying to
write to storage.

To avoid this issue, we always invoke wait() after killing the process.

Special code that do not want to wait for the child process should not
use this context manager.

Change-Id: Ida04e2c7ba092efdc2927ed9f460b0098ba2ad94
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65324
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>