build: Replace pep8 and pyflakes with flake8
flake8 is an extensible and configurable tool combining pep8, pyflakes
and other plugins.

This patch replace the pep8 and pyflakes targets with single flake8
target, and remove the complex and ugly tox.sh. Everyting is configured
in a simple way in tox.ini.

Since this version includes the newers versions of pyflakes and pep8, we are
ignoring some new errors and warnings. See tox.ini for the details.

To disable ignored warnings and errors, you can:

    tox -- --ignore=''

I enabled the --show-source option by default, so now we see exactly
where is the error:

    ./tests/functional/networkTests.py:465:29: W503 line break before binary operator
				<= (ARP_REQUEST_SIZE + DHCP_PACKET_SIZE),

I also enabled the --statistics option, showing:

    26    E123 closing bracket does not match indentation of opening bracket's line
    32    E226 missing whitespace around arithmetic operator
    45    E241 multiple spaces after ','
    31    E402 module level import not at top of file
    25    E731 do not assign a lambda expression, use a def
    24    W503 line break before binary operator

For more info see http://flake8.pycqa.org

Change-Id: Ib42430f045e4475c31cd370b4196214094c29cd1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65403
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>