build: Run the tests with tox
Run the tests via tox, installing the correct version of nose in a
virtual environment.  Since correct version is nose is always available,
the warning about nose version is not needed now.

By default, tox uses random PYTHONAHSHSEED on each run. This exposes bad
code assuming that unordered things are ordered. Turns out we have some
tests with this issue. To keep the build working, this option is
disabled now (PYTHONHASHSEED = 0). We should remove this option when the
bad tests are fixed.

Change-Id: I245a171940a5e869fd719a6410024ee77e8ad86c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65404
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>