xleases: Introduce the xlease module
This module manage external leases which are not attached to a volume.
These leases will be used for vm leases. External leases can also be
used for other scenarios when you don't have a volume or a volume is
on a storage not supporting leases.

External leases are stored in the xleases special volume. This module
manage allocation and querying of leases via a lease id. The module
maintains the mapping from lease id to lease offset.

Adding and removing leases will be performed only on the SPM, but once a
lease is created, any host can lookup the lease info and use it to
acquire the lease via sanlock.

Only index management is implemented. This module must also create the
sanlock leases in an atomic way.

Please check the feature page for more info:
http://www.ovirt.org/develop/release-management/features/storage/vm-leases/

Change-Id: I99b1fcff74240a568eee0b85c6f4beae84acbdb5
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>