virt: Set proper migration status in recovery
Recovery currently doesn't check for migrations and that may confuse
Engine.  For instance, a migrating HA VM on the source may end up in UP
status (instead of MIGRATION_SOURCE) after recovery and Engine may
interpret it as migration failure.  When the migration completes and
Engine sees the VM down on the source before finding it on the
destination, it may start it again on the source, resulting in split
brain.

So ignoring migrations in recovery is dangerous, even more with
post-copy migrations.  This patch tries to identify migration states
during recovery.  It's not a complete remedy, e.g. we don't start
migration source thread (so we must ensure that we get out of
MIGRATION_SOURCE state for that reason).  But we at least no longer lie
to Engine about the VM status.

The patch tries not to make things worse while making some things better,
without the ambition to make everything working perfectly.  After some
migrations that were recovered on Vdsm startup and then failed, it may happen
that the migration failure is not detected and the VM remains in the migrating
status on the source.  We might need better cooperation from libvirt (sending
proper events) or larger refactoring around the source thread to handle those,
supposedly rare, situations.  But just another Vdsm restart should be a good
enough remedy.  We may prepare more patches if it ever happens be a problem.

Change-Id: Ia7544e10a67f83b6409ef34713e966eb27539139
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>