vm: rename statsAge parameter
rename `statsAge' to `stats_age'.
Now it is pep8 friendlier, and it is easier to follow the code flow:
it is now possible to follow this flow end to end using
'git grep stats_age'.

It used to be called `statsAge' in camelCase only for legacy reasons.

Change-Id: If0b4601d0f789921c1c575f9ad19852d4ebecacc
Backport-To: 4.0
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1382578
Bug-Url: https://bugzilla.redhat.com/1382583
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65508
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65813