numa: Don't crash on hotplugged CPUs when retrieving NUMA info
getVmNumaNodeRuntimeInfo maps virtual (from Engine) nodes to physical
nodes by matching nodes and CPUs from the libvirt runtime and from the
Engine NUMA mapping.  CPUs that are not pinned to any node in the Engine
NUMA mapping (such as hotplugged CPUs) are irrelevant.  However the
matching cycle still looks for them in the Engine NUMA mapping
dictionary, resulting in KeyError.

This patch fixes the KeyError by simply skipping the CPUs not present in
the Engine NUMA mapping.

Change-Id: If7d00b897e430d9c4e8a23cb1451209b6f00e688
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65562
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI