vm: introduce a `monitorable' attribute
In the change I3e61626625a2e0517d55dc61e361f3f5eb690c00
we fixed the stats_age reporting, in order to correctly report
the real age of VM monitoring samples.

In the same change we acknowledged a possible change in behaviour:
depending on the timing of operations and on the load of the host,
VMs could be reported 'Unresponsive' for a little time while
starting up or shutting down.

This information is tecnhically correct: the monitoring subsystem
dutifully reports unknown sampling data, which is translated into
outdated values, thus unresponsive VM, but it is misleading for
both Engine and users.

There are two root causes here:
1. creation, destruction and monitoring of a VM are all operations
   which are asynchronous to each other, for both performance and
   historical reasons - and both reasons are still relevant today.
2. the semantic of this reporting (VM unresponsive for stats too old)
   is poorly defined, so the right thing is to keep backward
   compatibility, lacking better description.

To solve this issue we generalize the old VM.incomingMigrationPending
attribute into a 'monitorable' attribute.
A VM is monitorable if it is in a meaningful state, so after the
creation process is completed, or before the shutdown process is
initiated.

We should adjust responsiveness for stats too old
only if a VM is monitorable, not inconditionally.
Otherwise we can have misreports on slow startup
or slow shutdowns.

Please note that the `monitorable' state share similarities with
the VM state (VM.lastStatus attribute), but the two concepts
overlap only for a small part.
Lacking a precise definition, we use two independent variables
to track those attributes.

Change-Id: Idb12277a5f0fdaf680032af12fa7c104c3bd6ffb
Backport-To: 4.0
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1382578
Bug-Url: https://bugzilla.redhat.com/1382583
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65590
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>