tests: Require python-nose for automation
In commit 195937c2da59 (build: Disable tests during build) we removed
python-nose requirement, because we install nose using tox. Turns out
that we still need nose, since our testing infrastructure is not using
the virtual environment created by tox, and the nose version we install
with tox is actually unused.

The issue was probably hidden because we merged a patch without rebasing
in and running the tests again.

This patch is only the minimal fix to get the build working again, we
need more work to run the tests using tox virtual environment.

Change-Id: I8e6bfd4c1b06239b640baa13b72a2147ca664b07
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/65986
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>