threadPool: Name threadpool worker threads
Add required name argument to ThreadPool, and name the worker threads
name/N, similar to how executor threads are named.

The workers were changing the thread name to the task name, trying to
make it easier to follow tasks. Instead of renaming the thread for each
task, the logs when starting and ending a task were improved. Once task
is running, it is easy to follow the thread running the task until the
task finish or fail.

Change-Id: I5991718eb6316b40191eccb93a87ecf3ca69699a
Bug-Url: https://bugzilla.redhat.com/1141422
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/66109
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI