net: let users override Engine's default route
oVirt Engine sets default=True on the management network an on it alone.

Users may want to use another network for their host's default routing.

With this patch, they can modify this behavior by setting the custom
propery default_route to False on the management network attachement,
and to True on the would-be default route network.

Note that prior to using a custom property, one has to define it on
Engine by:
  sudo engine-config -g CustomDeviceProperties
  sudo engine-config -s CustomDeviceProperties='{type=interface;prop={default_reoute=^(true|false)$}}' --cver=4.0
and restart of ovirt-engine.

(Cf. https://bugzilla.redhat.com/show_bug.cgi?id=1200963#c42 )

Change-Id: Iaf392ea05e1e39acbf1b74a7a31acda9e750b36e
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/66127
Continuous-Integration: Jenkins CI
Reviewed-by: Germano Veit Michel <germano@redhat.com>
Reviewed-by: Edward Haas <edwardh@redhat.com>