net: save switch type and options while changing bond
In network/link/setup.py when we edit a bond and save it to running
config with partially changed slaves, we do not specify switch type.
network/netswitch.py:_split_switch_type_entries then can assume that
the bond is legacy.

This patch saves not only partially changed nics, but also used
switch type and options.

Change-Id: Id6e4df60a93639f7fe71de80a3cdb226d74d1948
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1379115