jsonrpc: introduce new client
This is a simple jsonrpc client which is not aware of the available
methods and parameters.

The user should consult the schema to construct the wanted command.

The client is invoked with:

   cli = client.connect(host, port, ssl)

For example:

   cli = client.connect('localhost', 54321, True)

Failure will result in ClientError exception.

Invoking commands:

    cli.call('Host.getVMList')

    cli.call('VM.getStats', {'vmID': 'bc26bd11-ee3b-4a56-80d4-770f383a47b9'})

The call method expects method name and a dictionary containing all mandatory
parameters, as they appear in the schema.

Complex parameters are also supported and share the same behavior.

Change-Id: Idd45d7e88bf2246beaf30550b12201917f32c354
Signed-off-by: Irit goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/64502
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>