API: Introduce getQemuImageInfo API.
The new API suppose to return information fetched using qemuimg info.
This should be usable to get the compat version of the volume.

This API will not use a job since it only returns information
about the volume and do not change it.
It is similar to how getVolumeInfo is working.

Change-Id: Ic170e2f142e8e70b534083c9468d249fe2478943
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/66295
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI