tests: Remove wrong crossImportTests.py
This tests is broken, promising to detect illegal imports in lib/vdsm,
but failing to detect wrong import such as "import API" in
vdsm.rpc.Bridge and vdsm.rpc.bindingxmlrpc. It also fail to detect
missing packages since it does not try to import any module from the
vdsm package.

This tests also complicates the build, having to run it separately from
the other tests, and this generate useless coverage noise in the CI.

Change-Id: Ic4f3c7d747656e44213cce19ce9c334311fab6c6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/66470
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>