containers: bootstrap the containers subpackage
The container support code, previously known with codename `convirt`,
adds another subpackage to Vdsm.
This subpackage will contain all the code needed to run docker images
as VMs. If the subpackage is not installed, no changes will be made
to Vdsm.

This patch bootstrap this subpackage shipping the supervdsm part.
Please note that this code is currently unused, a later patch will
make use of it.

We need this patch to complete the user-visible changes needed
by project `convirt`.

Change-Id: I08da0772bc85def0e6f9a3ea01a94fcf18545952
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/66646
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>