tox: Require specific version of flake8
Not requiring specific version of flake8 means we get the latest version
of flake8 and the latest version of the checkers (e.g. pycodestyle,
pyflakes), which means any update may break the build as new
style checks are added to the underlying tools.

I'm using flake8 3.0.4 (depending on pycodestyle 2.0.0) since this is
the version I have on Fedora 24, and the current ocde works fine with
it. Previously we got flake8 3.1.1 (depending on pycodestyle 2.1.1).

Upgrading flake8 must be done in a controlled way, checking the errors
exposed by each version and fixing them before upgrading the version.

Change-Id: If626f29be912e3b17fcfe9d0590c02027e92c4ec
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/66726
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI