sdm: Add set_volume_generation verb
In certain error scenarios we may want to update a volume's generation
ID before doing something else with the volume.  For example, we may
want to bump the generation before deleting a volume when we think there
may be an outstanding job related to this volume on an unresponsive
host.  Bumping the generation will effectively "fence" that outstanding
job since the job will fail when attempting to use a stale generation
value at any time in the future.

Change-Id: I9f5fba2efe640401e2a95cb3011abcd5967237b6
Signed-off-by: Adam Litke <alitke@redhat.com>