lvm: Add vdsm local lvm configuration
Add lvmlocal.conf with vdsm specific configuration:

- devices/global_filter - filter out ovirt lvs from all lvm commands on
  the host.

  This change is expected to prevent discovery and auto-activation of
  guest lvs on raw ovirt volumes. This problem is causing performance
  issues and may lead to data corruption.

  This change does not effect discovery and auto-activation of guest lvs
  on direct luns. There is no way vdsm can filter direct since we don't
  know which luns should be filtered.

- global/use_lvmetad - disable lvmetad usage in lvm commands

This first comments on the file contains vdsm metadata, used by vdsm
during upgrades. The default configuration looks like this:

   #REVISION: 1
   #PRIVATE: NO

With this configuration, vdsm will upgrade this file when running:

    vdsm-tool configure --module lvm

To control the local lvm configuration, the system administrator can
mark the file as private by modifying this line:

    #PRIVATE: YES

If the file is private, vdsm will never upgrade the file.

The vdsm metadata header starts at the first line and ends on the first
non-comment line.

This patch adds the default configuration.

Change-Id: I88fb6efef5084d4637ec0183598d220f96613e82
Bug-Url: https://bugzilla.redhat.com/1374545
Signed-off-by: Nir Soffer <nsoffer@redhat.com>