resourceManager: Use log.debug for unimportant messages
Registering an existing namespace is neither an interesting event nor
a state change that should be logged in info log level. This is
unfortunate debug message that we see each time a domain is refreshed,
because of the horrible way domain objects are managed.

This change removes these useless logs from the default log level:

    Resource namespace 01_img_93404146-0d02-4d98-8310-f356733d7176 already registered (sd:658)
    Resource namespace 02_vol_93404146-0d02-4d98-8310-f356733d7176 already registered (sd:667)
    Resource namespace 03_lvm_93404146-0d02-4d98-8310-f356733d7176 already registered (blockSD:874)

We need to check later if it is possible to avoid the registration in this
case, or maybe fix domain life cycle handling so we can register the namespace
after the previous instance unregistered it.

Change-Id: I785b411320735a2c24352da470f63a8b2dab554e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>