Adding simple client for sending gauge metrics to statsd port using udp
Introducing simple udp client. The client sends gauge metric statistics in
the metric 2.0 spec format - key:value|g.

Prior to this implementation we used python-statsd package which  is not built
out side epel. The code we used to use from python-statsd is only the part
that sends gauge by udp, we don't have any use for the rest currently.

Change-Id: I7bbc8a4c0e3afc25c205bfec58c5e77096803404
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>