utils: Don't keep reference to command
CommandStream receive a Popen instance and watch its stdout and stderr
streams. However it does not use the instance itself and there is no
reason to keep a reference to this object. Keeping reference prevent
garbage collection of the instance and the underling pipes, if some
other code is keeping reference to the CommandStream instance.

Change-Id: Ib376efe5861591ebc6f487356ad724f47db022ae
Signed-off-by: Nir Soffer <nsoffer@redhat.com>