protocoldetector: Handle socket.error in handle_accept
If socket.error is not handled in handle_accept, asyncore assumes that
the error is related to the listen socket and close it. However, a
socket.error may be raised by client socket if the remote side closed
the socket immediately, for example, during port scanning.

This is a regression introduced by commit b25dbdcce417 (Have the
protocol detector use the Reactor), replacing the previous
implementation with "better" asyncore base solution.

This patch fixes this issue by handling socket.error in handle_accept,
closing the client socket.

Change-Id: I10a41f6d3f07296ad42fbc92dbf66f4ec85530e0
Bug-Url: https://bugzilla.redhat.com/1391877
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/67056
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>