protocoldector: Handle fatal errors in acceptor
Every asyncore.dispatcher must implement handle_error, otherwise the
default implementation will be called, closing the dispatcher - even if
the error is not related to the dispatcher.

This patch implements handle_error logging the exception and closing the
socket if the error is related to the socket.

In handle_close, we panic instead of closing the dispatcher silently.
This is highly unexpected and the best thing we can do is die loudly.

Change-Id: I088a6b64772782cb8466141f1371fad87cb085ba
Signed-off-by: Nir Soffer <nsoffer@redhat.com>