v2v: handle requests to abort non-existing process
In situations when a child job fails to start properly (often due to
some internal vdsm problem) we would then raise an AttributeError
exception when trying to _abort() the job. This is because the _proc
attribute is None.

It is wiser not try to abort the process at all if there's nothing to
abort.

Change-Id: Ic34b2d10c6c9abbfa33d0d072c6633f03eb6091d
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>