storage: add discard argument to all related flows
This patch adds the possibility to pass a new argument, "discard", to
flows that might remove a logical volume.

Up until now, these flows got the argument "postZero" which caused vdsm
to post zeros on the lv right before removing it. From now and on, these
flows will also get the discard parameter, so that the lv can now be
discarded after it is zeroed (if postZero is true) and only then removed.

Note that the config value discard_enable that was added in patch
33af18b16 will still function, but the recommendation is to use the new
parameter rather than this config value as it will be removed in the
future.

Change-Id: I8e9bde2a0ad899ae56a18b4c7d6b405360d295f2
Bug-Url: https://bugzilla.redhat.com/1342919
Signed-off-by: Idan Shaby <ishaby@redhat.com>