net: NetworkManager python binding - Connection Settings
Introducing NM python binding through DBUS for VDSM.
Starting by exposing the connection settings interface, through which
it is possible to list and iterate through the available connections.

In NM terms, a connection is a configuration setting instance of a given
device.

The driver implementation uses dbus-python package (based on DBus-GLib).
It is an obsolete package, new applications should use pydbus, txdbus or
GDBus/QtDBus bindings.
Nevertheless, it has been chosen for the time being due to its
availability under most distros (RHEL7 as an example).

Change-Id: Ie5d38679e3c60d998980b310c61f34fb61bf394a
Signed-off-by: Edward Haas <edwardh@redhat.com>