Introduce sdm.api.amend_volume
Introduce the sdm amend opertion for a volume.

The operation amends the volume with specific parameters described
in the vol_attr.
Currently the only possible attribute that is allowed to be changed
is the volume qcow2 compat version.

The volume that is allowed to be amended is a qcow volume which is
not a SHARED volume (A template volume) and it is on a V4
storage domain.

Change-Id: I3e0b5e9f2d0c3742d915889c4ee2d2d9b5ce6a4f
Bug-Url: https://bugzilla.redhat.com/827529
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>