virt: Prettify domain and some other XMLs
etree formats XML on a single line, unless there are newlines in the parsed XML
(which is handled as text content by etree).  The compact form is not very
readable and there is apparently no way to make the output different with
etree.

This patch adds an option to make the XML output multiline and indented.
We use the already available manual formatting since it's probably not worth
using external libraries (such as bs4 or lxml) just for that and it seems to be
at least one order of magnitude faster than reformatting using minidom.

We use the pretty formatting at some places where minidom `toprettyxml' was
previously used: domain XML (to make it formatted in the log) and where passed
to hooks (in case someone assumes formatted XML in them).

Change-Id: I3666e0bd8eaf9a6b0f1728fd27ca0d28664649b3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>